sap/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dataclasses-json==0.6.3
Deprecated==1.2.14
fastapi==0.108.0
fireworks-ai==0.10.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
httpx-sse==0.3.1
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.352
langchain-cli==0.0.20
langchain-community==0.0.6
langchain-core==0.1.3
langchain-google-genai==0.0.5
langchainhub==0.1.14
langserve==0.0.37
langsmith==0.0.74
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
orjson==3.9.10
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyGithub==2.1.1
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
sse-starlette==1.8.2
starlette==0.32.0.post1
tenacity==8.2.3
tomlkit==0.12.3
tqdm==4.66.1
typer==0.9.0
types-requests==2.31.0.20231231
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.23.2
wrapt==1.16.0
yarl==1.9.4
